+++ demos/android-supabase-todolist/app/src/androidTest/java/com/example/androidexample/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.androidexample
+package com.powersync.androidexample
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.androidexample", appContext.packageName)
+        assertEquals("com.powersync.androidexample", appContext.packageName)
     }
 }
