+++ demos/android-supabase-todolist/app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.AndroidExample">
+            android:theme="@style/Theme.PowerSyncAndroidExample">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
