+++ demos/android-supabase-todolist/app/src/main/java/com/powersync/androidexample/MainActivity.kt
@@ -1,0 +1,19 @@
+package com.powersync.androidexample
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import com.powersync.demos.App
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        installSplashScreen()
+
+        super.onCreate(savedInstanceState)
+
+        setContent {
+            App()
+        }
+    }
+}
