+++ demos/android-supabase-todolist/app/src/main/java/com/powersync/androidexample/components/Menu.kt
