+++ demos/android-supabase-todolist/app/src/main/java/com/powersync/androidexample/powersync/Schema.kt
