+++ demos/android-supabase-todolist/app/src/main/java/com/powersync/androidexample/screens/SignUpScreen.kt
