+++ demos/android-supabase-todolist/app/src/main/java/com/powersync/androidexample/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.example.androidexample.ui.theme
+package com.powersync.androidexample.ui.theme
 
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
@@ -33,7 +33,7 @@ private val LightColorScheme = lightColorScheme(
 )
 
 @Composable
-fun AndroidExampleTheme(
+fun PowerSyncAndroidExampleTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     // Dynamic color is available on Android 12+
     dynamicColor: Boolean = true,
