+++ demos/android-supabase-todolist/app/src/main/java/com/powersync/androidexample/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.androidexample.ui.theme
+package com.powersync.androidexample.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
